requirements.txt
Django==1.7.1
MySQL-python==1.2.5
argparse==1.2.1
djangorestframework==3.2.3
psycopg2==2.6.1
wsgiref==0.1.2
